make copy and dpreview and download it
