+++ src/components/Success.js
@@ -1,12 +1,12 @@
 import React, { Component } from "react";
+import ReactLoading from 'react-loading';
 
 
 class Success extends Component {
-
   render() {
-
-    let urlSlide = `https://docs.google.com/presentation/d/${this.props.presentationId}/`
-    let urlDownload = `https://docs.google.com/presentation/d/${this.props.presentationId}/export/pptx`
+    let urlSlide = `https://docs.google.com/presentation/d/${this.props.copyId}/`
+    let urlDownload = `https://docs.google.com/presentation/d/${this.props.copyId}/export/pptx`
+    if(this.props.copyId !== ''){
     return (
       <div className="success-page container-fluid">
         <div className="success-page__icons">
@@ -15,16 +15,21 @@ class Success extends Component {
         </div>
         <div className="row d-flex justify-content-around">
           <div className="success-page__btn">
-          <a className="link-success" href={urlSlide}><button className="btn btn-outline-primary">View your presentation
-          </button></a>
+            <a className="link-success" href={urlSlide}><button className="btn btn-outline-primary">View your presentation
+            </button></a>
           </div>
           <div className="success-page__btn download-btn">
-          <a className="link-success" href= {urlDownload}
-         download="test.pptx"><button className="btn btn-outline-primary">Download your presentation</button></a>
+            <a className="link-success" href={urlDownload}
+              download="test.pptx"><button className="btn btn-outline-primary">Download your presentation</button></a>
           </div>
         </div>
       </div>
     );
+  } else{
+    return (
+    <ReactLoading type={'spinningBubbles'} color={'#990099'} height={100} width={100} />
+    )
+  }
   }
 }
